# Warn about unused whitelist options