I welcome suggestions on better ways to report warnings.

cc #1055