@sanxiyn 2018-12-19
Now I measured performance on internal codebase and I couldn't detect any change, but maybe our use cases are too small. Benchmark suggestions are welcome.

```
Use case #1

--whitelist-type 1
--whitelist-var 3

old
  time:   770.848 ms.   parse
  time:   632.237 ms.   compute_whitelisted_and_codegen_items
new
  time:   762.946 ms.   parse
  time:   629.872 ms.   compute_whitelisted_and_codegen_items
```

```
Use case #2

--whitelist-function 6
--whitelist-type 5
--whitelist-var 1

old
  time:   878.813 ms.   parse
  time:   743.636 ms.   compute_whitelisted_and_codegen_items
new
  time:   881.604 ms.   parse
  time:   726.318 ms.   compute_whitelisted_and_codegen_items
```