@emilio 2018-12-19
How many regexes are you using in your codebase? In Gecko we use quite a lot of them, see all the lists in:

  https://searchfox.org/mozilla-central/source/layout/style/ServoBindings.toml

I still think a configuration option for this would be useful, because it wouldn't be great to spam other developers about warnings that they don't know how to fix, or platform-dependent stuff.

It'd be great if you could do that but otherwise let me know and I'll try to take some time to make it happen as a followup to this PR before the next release.