@sanxiyn 2018-12-19
I already gave how many regexes we use. My understanding of RegexSet implementation is that slowdown happens only if multiple regexes actually match, and that doesn't seem to happen in Gecko, and I expect such cases to be rare.

I think it is very desirable to have this on by default, like Rust's unused warnings. I expect it won't be used in practice if it is not on by default. If it isn't too much trouble, can you run the benchmark for Gecko's use case?

I am uninterested in putting this behind a flag, so I appreciate your offer of a followup. Thanks.