@emilio 2018-12-23
Alright, this works for now, sorry for the lag (Christmas).

Will followup with the option, thank you for the patch! I agree this is useful :)